+++ src/Engines/MeilisearchEngine.php
@@ -4,12 +4,15 @@ namespace Laravel\Scout\Engines;
 
 use BackedEnum;
 use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Arr;
 use Illuminate\Support\LazyCollection;
 use Laravel\Scout\Builder;
 use Laravel\Scout\Contracts\UpdatesIndexSettings;
 use Laravel\Scout\Jobs\RemoveableScoutCollection;
 use Meilisearch\Client as MeilisearchClient;
 use Meilisearch\Contracts\IndexesQuery;
+use Meilisearch\Endpoints\Indexes;
+use Meilisearch\Exceptions\ApiException;
 use Meilisearch\Meilisearch;
 use Meilisearch\Search\SearchResult;
 
@@ -32,7 +35,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Create a new MeilisearchEngine instance.
      *
-     * @param  \Meilisearch\Client  $meilisearch
      * @param  bool  $softDelete
      * @return void
      */
@@ -136,8 +138,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Perform the given search on the engine.
      *
-     * @param  \Laravel\Scout\Builder  $builder
-     * @param  array  $searchParams
      * @return mixed
      */
     protected function performSearch(Builder $builder, array $searchParams = [])
@@ -174,7 +174,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Get the filter array for the query.
      *
-     * @param  \Laravel\Scout\Builder  $builder
      * @return string
      */
     protected function filters(Builder $builder)
@@ -227,9 +226,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
 
     /**
      * Get the sort array for the query.
-     *
-     * @param  \Laravel\Scout\Builder  $builder
-     * @return array
      */
     protected function buildSortFromOrderByClauses(Builder $builder): array
     {
@@ -248,7 +244,7 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
      */
     public function mapIds($results)
     {
-        if (0 === count($results['hits'])) {
+        if (count($results['hits']) === 0) {
             return collect();
         }
 
@@ -276,7 +272,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Get the results of the query as a Collection of primary keys.
      *
-     * @param  \Laravel\Scout\Builder  $builder
      * @return \Illuminate\Support\Collection
      */
     public function keys(Builder $builder)
@@ -289,14 +284,13 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Map the given results to instances of the given model.
      *
-     * @param  \Laravel\Scout\Builder  $builder
      * @param  mixed  $results
      * @param  \Illuminate\Database\Eloquent\Model  $model
      * @return \Illuminate\Database\Eloquent\Collection
      */
     public function map(Builder $builder, $results, $model)
     {
-        if (is_null($results) || 0 === count($results['hits'])) {
+        if (is_null($results) || count($results['hits']) === 0) {
             return $model->newCollection();
         }
 
@@ -326,7 +320,6 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
     /**
      * Map the given results to instances of the given model via a lazy collection.
      *
-     * @param  \Laravel\Scout\Builder  $builder
      * @param  mixed  $results
      * @param  \Illuminate\Database\Eloquent\Model  $model
      * @return \Illuminate\Support\LazyCollection
@@ -387,13 +380,22 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
      * Create a search index.
      *
      * @param  string  $name
-     * @param  array  $options
      * @return mixed
      *
      * @throws \Meilisearch\Exceptions\ApiException
      */
     public function createIndex($name, array $options = [])
     {
+        try {
+            $index = $this->meilisearch->getIndex($name);
+        } catch (ApiException $e) {
+            $index = null;
+        }
+
+        if ($index?->getUid() !== null) {
+            return $index;
+        }
+
         return $this->meilisearch->createIndex($name, $options);
     }
 
@@ -404,7 +406,13 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
      */
     public function updateIndexSettings($name, array $settings = [])
     {
-        $this->meilisearch->index($name)->updateSettings($settings);
+        $index = $this->meilisearch->index($name);
+
+        $index->updateSettings(Arr::except($settings, 'embedders'));
+
+        if (! empty($settings['embedders'])) {
+            $index->updateEmbedders($settings['embedders']);
+        }
     }
 
     /**
@@ -442,7 +450,7 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
         $tasks = [];
         $limit = 1000000;
 
-        $query = new IndexesQuery();
+        $query = new IndexesQuery;
         $query->setLimit($limit);
 
         $indexes = $this->meilisearch->getIndexes($query);