+++ src/Exceptions/ScoutException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Laravel\Scout\Exceptions;
+
+use Exception;
+
+class ScoutException extends Exception
+{
+    //
+}